regexp_nfa: cast int to long int
This matters only when DEBUG is defined:

```
gcc -c -I. -Iproto -DHAVE_CONFIG_H     -ggdb -O0 -DDEBUG -Wall -Wextra -Wshadow -Wmissing-prototypes -Wmaybe-uninitialized -Wno-cast-function-type -Wno-deprecated-declarations        -o objects/regexp.o regexp.c
In file included from regexp.c:2585:
regexp_nfa.c: In function ‘nfa_regmatch’:
regexp_nfa.c:6934:48: warning: format ‘%ld’ expects argument of type ‘long int’, but argument 2 has type ‘int’ [-Wformat=]
 6934 |       siemsg("INTERNAL: Negative state char: %ld", c);
      |                                              ~~^   ~
      |                                                |   |
      |                                                |   int
      |                                                long int
      |                                              %d
[...]
```

+++ src/regexp_nfa.c
@@ -6929,7 +6929,7 @@ nfa_regmatch(
 
 #ifdef DEBUG
 		if (c < 0)
-		    siemsg("INTERNAL: Negative state char: %ld", c);
+		    siemsg("INTERNAL: Negative state char: %ld", (long)c);
 #endif
 		result = (c == curc);
